Navigation rework (#56)
* Remove unnecessary async calls for navigation

* Refactor navigation bar rendering

* Create ChooseGoal action

* Delete stack class, replace all instances with array

* Begin changing navigation architecture to not store Components in store

* Create React component to hold a goal

* Create hash table for components

* Remove component hash table

Instead of using a hash table of components, we will create a new
component each time we want to render a goal. This will keep the data
and the display logic separate.

* Remove unnecessary components and write unit tests

* Remove some properties from Goal and BaseGoal, fix unit tests

The display and select properties were removed from Goal and BaseGoal,
because we do not want to store React components in the Redux store.
The steps property still uses React components, but that will be
fixed when we start implementing steps.

* Make goal id a string instead of a number

* Update goal constructors to reflect changes to User

* Fix LoginActions unit test

Axios was not being properly mocked. Note that when mocking, some
posts to the API are not being called. We know this is true, because
we tried logging to the console, and nothing was actually logged.
We are not sure why this is happening.

* Remove unused import in src/App/component.tsx

* Fix unit tests

* Make changes suggested in code review

* Make more changes suggested in code review

* Make changes based on code review

A comment was removed in GoalSwitcherActions.tsx because we are not
supposed to leave console logs in master