Make some words immutable
When entering data, the actions a user can perform are limited. They are
not allowed to edit words that already exist in the database. When a
word a user enters is flagged as a possible duplicate, the user can
choose to add the gloss they typed in as a new sense of the already
existing word. They can also choose to add the current semantic domain
to an already existing sense. In either case, the word from then on
becomes immutable.