Fix lgtm errors for c# (#380)
This PR should resolve all LGTM errors in the C# codebase. There are two types of errors fixed.

## User input used directly in path

There were many spots in the codebase where user input was used directly in paths this could be a security problem. Example:
if I call `v1/projects/{projectId}/words/download` with a purposefully malicious path like maybe `v1/projects/../../../important_stuff.txt/words/download` the user might be able to mess with files they shouldn't.

I have resolved this error by adding a function that confirms that a string is somekind of ID, specifically alphanumeric with dashes -. Any other character in the place of an ID will cause the request to fail.

Dereferenced variable that is always null
There was one amusing instance of a null check that then dereferences the value that is now confirmed to be null.