Move Tooltip inside RecorderIcon. (#839)
The ref property is not implicit with a functional component as it is for a traditional component. 
Thus, Tooltip cannot be used around a functional component, because it tries to pass its ref to its children. 
Previously in AudioRecorder, we had Tooltip wrapped around the functional component RecorderIcon. 
To fix the associated warnings, this moves Tooltip to within RecorderIcon. 
Should we ever need tooltip around a functional component: https://material-ui.com/guides/composition/#caveat-with-refs