Cleanup loadGoalData and blacklist building. (#902)
Removes from loadGoalData() the block on words that were seen in a singleton set or a blacklist set (which had prevented them from being compared to other words). 
The latter block was because getNextDups() has some sets that are strict subsets of previous sets. If you blacklisted a set, you shouldn't have to consider its subsets. To target that problem, this pr updates the blacklisting function to blacklist not just a specified set but also its subsets.