Implement cert-client mode for the certmgr container (#1077)
* Add is_reachable to utility functions in utils.py

* Change aws_s3_get to return True if successful

* Implement the cert-client mode for the certmgr container

* Configure nuc group to use certmgr in cert-client mode

* Fix call to get_time_to_expire

* Add module docstring

* Merge remote-tracking branch 'origin/master' into certmgr_proxy_client

* Fix member variable type; documentation.

* Improve certmgr logging

* Add procps package

* Replace sleep with sigtimedwait

* Merge branch 'master' into certmgr_proxy_client

* Merge branch 'certmgr_proxy_client' into detect_ethernet

* Add trigger to renew certificate

Trigger a check for an updated certificate when the ethernet cable is 
plugged in, i.e. there is a "routable" event for networkd-dispacther.  
This is only installed when the cert_mode is set to "cert-client"

* Remove unnecessary log messages

* Code tidying.

Rename variable
Improve comment

* Remove cast function calls

Remove cast function calls when calling get_settings - get_settings will 
always return a str and type conversion is performed as necessary.

In addition, the calling parameters are added to the SIGUSR1 handler 
with their types.

* Change handler to print signal number; owner is root

* Address review comments

* Remove superfluous type hinting