Remove NUC instruction pages from frontend image
A user needs to connect to the NUC using the NUC's WiFi access point.  If the user is connected to the internet,
they will be connected to the production server.  The NUC instruction pages direct the user to connect to the WiFi access point and reload.  These pages are no longer served by the frontend - the cert-proxy-server is now resposible for providing these pages.