Update spell-check dictionaries (#3717)
The dictionary for Hindi (hi) has grown a lot since this was last done, so dropped the custom threshold adjustment (changing from -t 1000 to the default -t 2000) and added a max word-length (-m 6) to keep the updated src/resources/dictionaries/hi down to 2.4 MB.

The Spanish (es) dictionary has shrunk a bit, so added a custom threshold (from default -T 20000 to -T 15000) to maintain some of the file splitting.

Dropped numeral support, to clear noisy entries, especially in the French (fr) wordlist.

Switched from NFKD to NFD normalization to allow for unexpected characters and to match #3718.